[bugfix] Fix bug in binding
